HOME-311 - Code cleanup
