--- smarthome/constants/services.swift
@@ -1,10 +1,0 @@
-//
-//  api.swift
-//  smarthome
-//
-//  Created by Oskar on 16/01/2021.
-//
-
-struct CServices {
-    static let service = "http://93.180.188.173:3222"
-}
