+++ smarthome/models/userModel.swift
@@ -9,3 +9,7 @@ struct UserModel {
     var username: String
     var password: String
 }
+
+struct UserSessionResponse: Codable {
+    var isSession: Bool
+}
