+++ smarthome/repositories/userRepository.swift
@@ -1,12 +1,19 @@
+
 //
-//  userRepository.swift
+//  UserRespository.swift
 //  smarthome
 //
-//  Created by Oskar on 16/01/2021.
+//  Created by Oskar on 27/01/2021.
 //
-
-import Foundation
 
 struct UserRepository {
     let api: Api
+    
+    func login(username: String, apiAddress: String, password: String, completion: @escaping (Bool) -> ()) {
+        self.api.post(url: "\(apiAddress)/api/login",
+                      body: "username=\(username)&password=\(password)",
+                      completion: { (data: UserSessionResponse?) in
+                        completion(data?.isSession ?? false)
+                      })
+    }
 }
