+++ smarthome/services/api.swift
@@ -7,40 +7,30 @@
 
 import Foundation
 
-struct UserSession: Codable {
-    var isSession: Bool
-}
-struct Embedded: Codable {
-    var agents: Array<AgentModel>
-}
-struct AgentsResponse: Codable {
-    var _embedded: Embedded
-}
-
 struct Api {
-    func get(url: String, completion: @escaping (Array<AgentModel>) -> ()) {
+    func get<T: Codable>(url: String, type: T.Type, completion: @escaping (T?) -> ()) {
         let urlTarget = URL(string: url)!
         let request = URLRequest(url: urlTarget)
         
         let task = URLSession.shared.dataTask(with: request) {(data, response, error) in
             guard let data = data else {
-                completion([])
+                completion(nil)
                 return
             }
             
             do {
-                let agentResponse = try JSONDecoder().decode(AgentsResponse.self, from: data)
-                completion(agentResponse._embedded.agents)
+                let response = try JSONDecoder().decode(T.self, from: data)
+                completion(response)
                 return
             } catch {
-                completion([])
+                completion(nil)
             }
         }
         
         task.resume()
     }
     
-    func post(url: String, body: String, completion: @escaping (Bool) -> ()) {
+    func post<T: Codable>(url: String, body: String, completion: @escaping (T?) -> ()) {
         let urlTarget = URL(string: url)!
         
         var request = URLRequest(url: urlTarget)
@@ -49,16 +39,16 @@ struct Api {
         
         let task = URLSession.shared.dataTask(with: request) {(data, response, error) in
             guard let data = data else {
-                completion(false)
+                completion(nil)
                 return
             }
             
             do {
-                let userSession = try JSONDecoder().decode(UserSession.self, from: data)
-                completion(userSession.isSession)
+                let response = try JSONDecoder().decode(T.self, from: data)
+                completion(response)
                 return
             } catch {
-                completion(false)
+                completion(nil)
             }
         }
